// src/app/common/components/chartarea/chart/model/state/selectedmeasure/types/AggregationFunction.ts
export type AggregationFunction =
  | 'NONE'
  | 'SUM'
  | 'AVG'
  | 'COUNT'
  | 'DISTINCT'
  | 'MIN'
  | 'MAX'
  | 'STDDEV'
  | 'VAR'
  | 'LATEST'
  | 'MEDIAN'
  | '5TH PCTL'
  | '25TH PCTL'
  | '75TH PCTL'
  | '95TH PCTL'
  | 'INCREASE'
  | 'DECREASE'
  | 'LAST INCR'
  | 'LAST DECR'
  | 'INSTANT INCR'
  | 'INSTANT DECR'
  | 'MAX INCR'
  | 'MAX DECR';